moved level select to new cpp
added header
