+++ fifteen_game.h
@@ -2,3 +2,4 @@
 #include "Project_window.h"
 #include "End_screen.h"
 #include "Game_screen.h"
+#include "Level_select.h"
